pyproject.toml
[project]
name = "corporate-profilex"
version = "0.1.0"
description = "GenAI-powered corporate profiling platform with multi-agent scraping and insights."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.2",
    "black>=25.11.0",
    "chromadb>=1.3.5",
    "duckdb>=1.4.2",
    "faiss-cpu>=1.13.0",
    "google-search-results==2.0.0",
    "gradio>=6.0.1",
    "ipython>=9.7.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "pytest>=9.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["agents", "core", "storage", "ui", "ingestion", "diagrams"]
